+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/headless-example
 
+## 0.3.113
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.91
+  - @ballerine/workflow-browser-sdk@0.6.114
+
+## 0.3.112
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.90
+  - @ballerine/workflow-browser-sdk@0.6.113
+
 ## 0.3.111
 
 ### Patch Changes